Tag management + cleanup.
Improved docker image tag management - there is a new script that will
find NAV ProductVersion and this will automatically appear as a docker
image tag.

Cleanup of the the _*.bat files - there is only one right now that will
preconfigure all required variables and this _presetvars.bat is being
called from the rest of the files.